[Refactor]split column reader make it more readable
