+++ be/src/formats/CMakeLists.txt
@@ -53,6 +53,8 @@ add_library(Formats STATIC
         parquet/column_chunk_reader.cpp
         parquet/column_converter.cpp
         parquet/column_reader.cpp
+        parquet/scalar_column_reader.cpp
+        parquet/complex_column_reader.cpp
         parquet/encoding.cpp
         parquet/level_codec.cpp
         parquet/page_reader.cpp
