+++ be/src/formats/parquet/column_reader.cpp
@@ -24,402 +24,22 @@
 #include <unordered_map>
 #include <utility>
 
-#include "column/array_column.h"
 #include "column/chunk.h"
 #include "column/column_helper.h"
-#include "column/fixed_length_column.h"
-#include "column/map_column.h"
 #include "column/nullable_column.h"
-#include "column/struct_column.h"
 #include "common/compiler_util.h"
 #include "exec/exec_node.h"
 #include "exec/hdfs_scanner.h"
 #include "exprs/expr_context.h"
-#include "formats/parquet/column_converter.h"
-#include "formats/parquet/page_index_reader.h"
-#include "formats/parquet/schema.h"
-#include "formats/parquet/stored_column_reader.h"
-#include "formats/parquet/stored_column_reader_with_index.h"
-#include "fs/fs.h"
+#include "formats/parquet/complex_column_reader.h"
+#include "formats/parquet/scalar_column_reader.h"
 #include "gen_cpp/Descriptors_types.h"
 #include "gen_cpp/parquet_types.h"
-#include "gutil/casts.h"
-#include "gutil/strings/substitute.h"
-#include "io/shared_buffered_input_stream.h"
-#include "runtime/types.h"
 #include "simd/batch_run_counter.h"
-#include "simd/simd.h"
 #include "storage/column_or_predicate.h"
 #include "storage/column_predicate.h"
-#include "storage/types.h"
-#include "util/runtime_profile.h"
-#include "util/slice.h"
-#include "util/stopwatch.hpp"
-#include "util/thrift_util.h"
-#include "utils.h"
 
 namespace starrocks::parquet {
-
-template <typename TOffset, typename TIsNull>
-static void def_rep_to_offset(const LevelInfo& level_info, const level_t* def_levels, const level_t* rep_levels,
-                              size_t num_levels, TOffset* offsets, TIsNull* is_nulls, size_t* num_offsets,
-                              bool* has_null) {
-    size_t offset_pos = 0;
-    for (int i = 0; i < num_levels; ++i) {
-        // when def_level is less than immediate_repeated_ancestor_def_level, it means that level
-        // will affect its ancestor.
-        // when rep_level is greater than max_rep_level, this means that level affects its
-        // descendants.
-        // So we can skip this levels
-        if (def_levels[i] < level_info.immediate_repeated_ancestor_def_level ||
-            rep_levels[i] > level_info.max_rep_level) {
-            continue;
-        }
-        if (rep_levels[i] == level_info.max_rep_level) {
-            offsets[offset_pos]++;
-            continue;
-        }
-
-        // Start for a new row
-        offset_pos++;
-        offsets[offset_pos] = offsets[offset_pos - 1];
-        if (def_levels[i] >= level_info.max_def_level) {
-            offsets[offset_pos]++;
-        }
-
-        // when def_level equals with max_def_level, this is a non null element or a required element
-        // when def_level equals with (max_def_level - 1), this indicates an empty array
-        // when def_level less than (max_def_level - 1) it means this array is null
-        if (def_levels[i] >= level_info.max_def_level - 1) {
-            is_nulls[offset_pos - 1] = 0;
-        } else {
-            is_nulls[offset_pos - 1] = 1;
-            *has_null = true;
-        }
-    }
-    *num_offsets = offset_pos;
-}
-
-class ScalarColumnReader : public ColumnReader {
-public:
-    explicit ScalarColumnReader(const ColumnReaderOptions& opts) : _opts(opts) {}
-    ~ScalarColumnReader() override = default;
-
-    Status init(const ParquetField* field, const TypeDescriptor& col_type,
-                const tparquet::ColumnChunk* chunk_metadata) {
-        _field = field;
-        _col_type = &col_type;
-        _chunk_metadata = chunk_metadata;
-        RETURN_IF_ERROR(ColumnConverterFactory::create_converter(*field, col_type, _opts.timezone, &converter));
-        return StoredColumnReader::create(_opts, field, chunk_metadata, &_reader);
-    }
-
-    Status read_range(const Range<uint64_t>& range, const Filter* filter, ColumnPtr& dst) override {
-        DCHECK(_field->is_nullable ? dst->is_nullable() : true);
-        _need_lazy_decode =
-                _dict_filter_ctx != nullptr || (_can_lazy_decode && filter != nullptr &&
-                                                SIMD::count_nonzero(*filter) * 1.0 / filter->size() < FILTER_RATIO);
-        ColumnContentType content_type = !_need_lazy_decode ? ColumnContentType::VALUE : ColumnContentType::DICT_CODE;
-        if (_need_lazy_decode) {
-            if (_dict_code == nullptr) {
-                _dict_code = ColumnHelper::create_column(
-                        TypeDescriptor::from_logical_type(ColumnDictFilterContext::kDictCodePrimitiveType), true);
-            }
-            dst = _dict_code;
-            dst->reserve(range.span_size());
-        }
-        if (!converter->need_convert) {
-            SCOPED_RAW_TIMER(&_opts.stats->column_read_ns);
-            return _reader->read_range(range, filter, content_type, dst.get());
-        } else {
-            auto column = converter->create_src_column();
-            {
-                SCOPED_RAW_TIMER(&_opts.stats->column_read_ns);
-                RETURN_IF_ERROR(_reader->read_range(range, filter, content_type, column.get()));
-            }
-            SCOPED_RAW_TIMER(&_opts.stats->column_convert_ns);
-            return converter->convert(column, dst.get());
-        }
-    }
-
-    void get_levels(level_t** def_levels, level_t** rep_levels, size_t* num_levels) override {
-        _reader->get_levels(def_levels, rep_levels, num_levels);
-    }
-
-    Status get_dict_values(const std::vector<int32_t>& dict_codes, const NullableColumn& nulls,
-                           Column* column) override {
-        return _reader->get_dict_values(dict_codes, nulls, column);
-    }
-
-    void set_need_parse_levels(bool need_parse_levels) override { _reader->set_need_parse_levels(need_parse_levels); }
-
-    bool try_to_use_dict_filter(ExprContext* ctx, bool is_decode_needed, const SlotId slotId,
-                                const std::vector<std::string>& sub_field_path, const size_t& layer) override {
-        if (sub_field_path.size() != layer) {
-            return false;
-        }
-
-        if (!_col_type->is_string_type()) {
-            return false;
-        }
-
-        if (_column_all_pages_dict_encoded()) {
-            if (_dict_filter_ctx == nullptr) {
-                _dict_filter_ctx = std::make_unique<ColumnDictFilterContext>();
-                _dict_filter_ctx->is_decode_needed = is_decode_needed;
-                _dict_filter_ctx->sub_field_path = sub_field_path;
-                _dict_filter_ctx->slot_id = slotId;
-            }
-            _dict_filter_ctx->conjunct_ctxs.push_back(ctx);
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    Status rewrite_conjunct_ctxs_to_predicate(bool* is_group_filtered, const std::vector<std::string>& sub_field_path,
-                                              const size_t& layer) override {
-        DCHECK_EQ(sub_field_path.size(), layer);
-        return _dict_filter_ctx->rewrite_conjunct_ctxs_to_predicate(_reader.get(), is_group_filtered);
-    }
-
-    void set_can_lazy_decode(bool can_lazy_decode) override {
-        _can_lazy_decode = can_lazy_decode && _col_type->is_string_type() && _column_all_pages_dict_encoded();
-    }
-
-    Status filter_dict_column(const ColumnPtr& column, Filter* filter, const std::vector<std::string>& sub_field_path,
-                              const size_t& layer) override {
-        DCHECK_EQ(sub_field_path.size(), layer);
-        return _dict_filter_ctx->predicate->evaluate_and(column.get(), filter->data());
-    }
-
-    Status fill_dst_column(ColumnPtr& dst, const ColumnPtr& src) override {
-        if (!_need_lazy_decode) {
-            dst->swap_column(*src);
-        } else {
-            if (_dict_filter_ctx == nullptr || _dict_filter_ctx->is_decode_needed) {
-                ColumnPtr& dict_values = dst;
-                dict_values->reserve(src->size());
-
-                // decode dict code to dict values.
-                // note that in dict code, there could be null value.
-                const ColumnPtr& dict_codes = src;
-                auto* codes_nullable_column = ColumnHelper::as_raw_column<NullableColumn>(dict_codes);
-                auto* codes_column =
-                        ColumnHelper::as_raw_column<FixedLengthColumn<int32_t>>(codes_nullable_column->data_column());
-                RETURN_IF_ERROR(get_dict_values(codes_column->get_data(), *codes_nullable_column, dict_values.get()));
-                DCHECK_EQ(dict_codes->size(), dict_values->size());
-                if (dict_values->is_nullable()) {
-                    auto* nullable_codes = down_cast<NullableColumn*>(dict_codes.get());
-                    auto* nullable_values = down_cast<NullableColumn*>(dict_values.get());
-                    nullable_values->null_column_data().swap(nullable_codes->null_column_data());
-                    nullable_values->set_has_null(nullable_codes->has_null());
-                }
-            } else {
-                dst->append_default(src->size());
-            }
-
-            src->reset_column();
-        }
-        return Status::OK();
-    }
-
-    void collect_column_io_range(std::vector<io::SharedBufferedInputStream::IORange>* ranges, int64_t* end_offset,
-                                 ColumnIOType type, bool active) override;
-
-    const tparquet::ColumnChunk* get_chunk_metadata() override { return _chunk_metadata; }
-
-    const ParquetField* get_column_parquet_field() override { return _field; }
-
-    StatusOr<tparquet::OffsetIndex*> get_offset_index(const uint64_t rg_first_row) override {
-        if (_offset_index_ctx == nullptr) {
-            _offset_index_ctx = std::make_unique<ColumnOffsetIndexCtx>();
-            _offset_index_ctx->rg_first_row = rg_first_row;
-            int64_t offset_index_offset = _chunk_metadata->offset_index_offset;
-            uint32_t offset_index_length = _chunk_metadata->offset_index_length;
-            std::vector<uint8_t> offset_index_data;
-            offset_index_data.reserve(offset_index_length);
-            RETURN_IF_ERROR(
-                    _opts.file->read_at_fully(offset_index_offset, offset_index_data.data(), offset_index_length));
-
-            RETURN_IF_ERROR(deserialize_thrift_msg(offset_index_data.data(), &offset_index_length,
-                                                   TProtocolType::COMPACT, &_offset_index_ctx->offset_index));
-        }
-        return &_offset_index_ctx->offset_index;
-    }
-
-    void select_offset_index(const SparseRange<uint64_t>& range, const uint64_t rg_first_row) override;
-
-private:
-    // Returns true if all of the data pages in the column chunk are dict encoded
-    bool _column_all_pages_dict_encoded();
-
-    const ColumnReaderOptions& _opts;
-
-    std::unique_ptr<StoredColumnReader> _reader;
-    const ParquetField* _field = nullptr;
-
-    std::unique_ptr<ColumnDictFilterContext> _dict_filter_ctx;
-    const TypeDescriptor* _col_type = nullptr;
-    const tparquet::ColumnChunk* _chunk_metadata = nullptr;
-    std::unique_ptr<ColumnOffsetIndexCtx> _offset_index_ctx;
-
-    // _can_lazy_decode means string type and all page dict code
-    bool _can_lazy_decode = false;
-    // we use lazy decode adaptively because of RLE && decoder may be better than filter && decoder
-    static constexpr double FILTER_RATIO = 0.2;
-    bool _need_lazy_decode = false;
-    // dict code
-    ColumnPtr _dict_code = nullptr;
-};
-
-bool ScalarColumnReader::_column_all_pages_dict_encoded() {
-    // The Parquet spec allows for column chunks to have mixed encodings
-    // where some data pages are dictionary-encoded and others are plain
-    // encoded. For example, a Parquet file writer might start writing
-    // a column chunk as dictionary encoded, but it will switch to plain
-    // encoding if the dictionary grows too large.
-    //
-    // In order for dictionary filters to skip the entire row group,
-    // the conjuncts must be evaluated on column chunks that are entirely
-    // encoded with the dictionary encoding. There are two checks
-    // available to verify this:
-    // 1. The encoding_stats field on the column chunk metadata provides
-    //    information about the number of data pages written in each
-    //    format. This allows for a specific check of whether all the
-    //    data pages are dictionary encoded.
-    // 2. The encodings field on the column chunk metadata lists the
-    //    encodings used. If this list contains the dictionary encoding
-    //    and does not include unexpected encodings (i.e. encodings not
-    //    associated with definition/repetition levels), then it is entirely
-    //    dictionary encoded.
-    const tparquet::ColumnMetaData& column_metadata =
-            _opts.row_group_meta->columns[_field->physical_column_index].meta_data;
-    if (column_metadata.__isset.encoding_stats) {
-        // Condition #1 above
-        for (const tparquet::PageEncodingStats& enc_stat : column_metadata.encoding_stats) {
-            if (enc_stat.page_type == tparquet::PageType::DATA_PAGE &&
-                (enc_stat.encoding != tparquet::Encoding::PLAIN_DICTIONARY &&
-                 enc_stat.encoding != tparquet::Encoding::RLE_DICTIONARY) &&
-                enc_stat.count > 0) {
-                return false;
-            }
-        }
-    } else {
-        // Condition #2 above
-        bool has_dict_encoding = false;
-        bool has_nondict_encoding = false;
-        for (const tparquet::Encoding::type& encoding : column_metadata.encodings) {
-            if (encoding == tparquet::Encoding::PLAIN_DICTIONARY || encoding == tparquet::Encoding::RLE_DICTIONARY) {
-                has_dict_encoding = true;
-            }
-
-            // RLE and BIT_PACKED are used for repetition/definition levels
-            if (encoding != tparquet::Encoding::PLAIN_DICTIONARY && encoding != tparquet::Encoding::RLE_DICTIONARY &&
-                encoding != tparquet::Encoding::RLE && encoding != tparquet::Encoding::BIT_PACKED) {
-                has_nondict_encoding = true;
-                break;
-            }
-        }
-        // Not entirely dictionary encoded if:
-        // 1. No dictionary encoding listed
-        // OR
-        // 2. Some non-dictionary encoding is listed
-        if (!has_dict_encoding || has_nondict_encoding) {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-void ScalarColumnReader::collect_column_io_range(std::vector<io::SharedBufferedInputStream::IORange>* ranges,
-                                                 int64_t* end_offset, ColumnIOType type, bool active) {
-    const auto& column = _opts.row_group_meta->columns[_field->physical_column_index];
-    if (type == ColumnIOType::PAGES) {
-        const tparquet::ColumnMetaData& column_metadata = column.meta_data;
-        if (_offset_index_ctx != nullptr) {
-            // add dict page
-            if (column_metadata.__isset.dictionary_page_offset) {
-                auto r = io::SharedBufferedInputStream::IORange(
-                        column_metadata.dictionary_page_offset,
-                        column_metadata.data_page_offset - column_metadata.dictionary_page_offset, active);
-                ranges->emplace_back(r);
-                *end_offset = std::max(*end_offset, r.offset + r.size);
-            }
-            _offset_index_ctx->collect_io_range(ranges, end_offset, active);
-        } else {
-            int64_t offset = 0;
-            if (column_metadata.__isset.dictionary_page_offset) {
-                offset = column_metadata.dictionary_page_offset;
-            } else {
-                offset = column_metadata.data_page_offset;
-            }
-            int64_t size = column_metadata.total_compressed_size;
-            auto r = io::SharedBufferedInputStream::IORange(offset, size, active);
-            ranges->emplace_back(r);
-            *end_offset = std::max(*end_offset, offset + size);
-        }
-    } else if (type == ColumnIOType::PAGE_INDEX) {
-        // only active column need column index
-        if (column.__isset.column_index_offset && active) {
-            auto r = io::SharedBufferedInputStream::IORange(column.column_index_offset, column.column_index_length);
-            ranges->emplace_back(r);
-        }
-        // all column need offset index
-        if (column.__isset.offset_index_offset) {
-            auto r = io::SharedBufferedInputStream::IORange(column.offset_index_offset, column.offset_index_length);
-            ranges->emplace_back(r);
-        }
-    }
-}
-
-void ScalarColumnReader::select_offset_index(const SparseRange<uint64_t>& range, const uint64_t rg_first_row) {
-    if (_offset_index_ctx == nullptr) {
-        if (!_chunk_metadata->__isset.offset_index_offset) {
-            return;
-        }
-        auto st = get_offset_index(rg_first_row);
-        if (!st.ok()) {
-            return;
-        }
-    }
-    size_t page_num = _offset_index_ctx->offset_index.page_locations.size();
-    size_t range_size = range.size();
-
-    size_t range_idx = 0;
-    Range<uint64_t> r = range[range_idx++];
-
-    for (size_t i = 0; i < page_num; i++) {
-        int64_t first_row = _offset_index_ctx->offset_index.page_locations[i].first_row_index + rg_first_row;
-        int64_t end_row = first_row;
-        if (i != page_num - 1) {
-            end_row = _offset_index_ctx->offset_index.page_locations[i + 1].first_row_index + rg_first_row;
-        } else {
-            // a little trick, we don't care about the real rows of the last page.
-            if (range.end() < first_row) {
-                _offset_index_ctx->page_selected.emplace_back(false);
-                continue;
-            } else {
-                end_row = range.end();
-            }
-        }
-        if (end_row <= r.begin()) {
-            _offset_index_ctx->page_selected.emplace_back(false);
-            continue;
-        }
-        while (first_row >= r.end() && range_idx < range_size) {
-            r = range[range_idx++];
-        }
-        _offset_index_ctx->page_selected.emplace_back(first_row < r.end() && end_row > r.begin());
-    }
-    const tparquet::ColumnMetaData& column_metadata =
-            _opts.row_group_meta->columns[_field->physical_column_index].meta_data;
-    bool has_dict_page = column_metadata.__isset.dictionary_page_offset;
-    // be compatible with PARQUET-1850
-    has_dict_page |= _offset_index_ctx->check_dictionary_page(column_metadata.data_page_offset);
-    _reader = std::make_unique<StoredColumnReaderWithIndex>(std::move(_reader), _offset_index_ctx.get(), has_dict_page);
-}
 
 Status ColumnDictFilterContext::rewrite_conjunct_ctxs_to_predicate(StoredColumnReader* reader,
                                                                    bool* is_group_filtered) {
@@ -517,480 +137,6 @@ Status ColumnDictFilterContext::rewrite_conjunct_ctxs_to_predicate(StoredColumnR
 
     return Status::OK();
 }
-
-class ListColumnReader : public ColumnReader {
-public:
-    explicit ListColumnReader(const ColumnReaderOptions& opts) {}
-    ~ListColumnReader() override = default;
-
-    Status init(const ParquetField* field, std::unique_ptr<ColumnReader> element_reader) {
-        _field = field;
-        _element_reader = std::move(element_reader);
-        return Status::OK();
-    }
-
-    Status read_range(const Range<uint64_t>& range, const Filter* filter, ColumnPtr& dst) override {
-        NullableColumn* nullable_column = nullptr;
-        ArrayColumn* array_column = nullptr;
-        if (dst->is_nullable()) {
-            nullable_column = down_cast<NullableColumn*>(dst.get());
-            DCHECK(nullable_column->mutable_data_column()->is_array());
-            array_column = down_cast<ArrayColumn*>(nullable_column->mutable_data_column());
-        } else {
-            DCHECK(dst->is_array());
-            DCHECK(!_field->is_nullable);
-            array_column = down_cast<ArrayColumn*>(dst.get());
-        }
-        auto& child_column = array_column->elements_column();
-        RETURN_IF_ERROR(_element_reader->read_range(range, filter, child_column));
-
-        level_t* def_levels = nullptr;
-        level_t* rep_levels = nullptr;
-        size_t num_levels = 0;
-        _element_reader->get_levels(&def_levels, &rep_levels, &num_levels);
-
-        auto& offsets = array_column->offsets_column()->get_data();
-        offsets.resize(num_levels + 1);
-        NullColumn null_column(num_levels);
-        auto& is_nulls = null_column.get_data();
-        size_t num_offsets = 0;
-        bool has_null = false;
-        def_rep_to_offset(_field->level_info, def_levels, rep_levels, num_levels, &offsets[0], &is_nulls[0],
-                          &num_offsets, &has_null);
-        offsets.resize(num_offsets + 1);
-        is_nulls.resize(num_offsets);
-
-        if (dst->is_nullable()) {
-            DCHECK(nullable_column != nullptr);
-            nullable_column->mutable_null_column()->swap_column(null_column);
-            nullable_column->set_has_null(has_null);
-        }
-
-        return Status::OK();
-    }
-
-    void get_levels(level_t** def_levels, level_t** rep_levels, size_t* num_levels) override {
-        _element_reader->get_levels(def_levels, rep_levels, num_levels);
-    }
-
-    void set_need_parse_levels(bool need_parse_levels) override {
-        _element_reader->set_need_parse_levels(need_parse_levels);
-    }
-
-    void collect_column_io_range(std::vector<io::SharedBufferedInputStream::IORange>* ranges, int64_t* end_offset,
-                                 ColumnIOType type, bool active) override {
-        _element_reader->collect_column_io_range(ranges, end_offset, type, active);
-    }
-
-    void select_offset_index(const SparseRange<uint64_t>& range, const uint64_t rg_first_row) override {
-        _element_reader->select_offset_index(range, rg_first_row);
-    }
-
-private:
-    const ParquetField* _field = nullptr;
-    std::unique_ptr<ColumnReader> _element_reader;
-};
-
-class MapColumnReader : public ColumnReader {
-public:
-    explicit MapColumnReader() = default;
-    ~MapColumnReader() override = default;
-
-    Status init(const ParquetField* field, std::unique_ptr<ColumnReader> key_reader,
-                std::unique_ptr<ColumnReader> value_reader) {
-        _field = field;
-        _key_reader = std::move(key_reader);
-        _value_reader = std::move(value_reader);
-
-        // Check must has one valid column reader
-        if (_key_reader == nullptr && _value_reader == nullptr) {
-            return Status::InternalError("No avaliable parquet subfield column reader in MapColumn");
-        }
-
-        return Status::OK();
-    }
-
-    Status read_range(const Range<uint64_t>& range, const Filter* filter, ColumnPtr& dst) override {
-        NullableColumn* nullable_column = nullptr;
-        MapColumn* map_column = nullptr;
-        if (dst->is_nullable()) {
-            nullable_column = down_cast<NullableColumn*>(dst.get());
-            DCHECK(nullable_column->mutable_data_column()->is_map());
-            map_column = down_cast<MapColumn*>(nullable_column->mutable_data_column());
-        } else {
-            DCHECK(dst->is_map());
-            DCHECK(!_field->is_nullable);
-            map_column = down_cast<MapColumn*>(dst.get());
-        }
-        auto& key_column = map_column->keys_column();
-        auto& value_column = map_column->values_column();
-        if (_key_reader != nullptr) {
-            RETURN_IF_ERROR(_key_reader->read_range(range, filter, key_column));
-        }
-
-        if (_value_reader != nullptr) {
-            RETURN_IF_ERROR(_value_reader->read_range(range, filter, value_column));
-        }
-
-        // if neither key_reader not value_reader is nullptr , check the value_column size is the same with key_column
-        DCHECK((_key_reader == nullptr) || (_value_reader == nullptr) || (value_column->size() == key_column->size()));
-
-        level_t* def_levels = nullptr;
-        level_t* rep_levels = nullptr;
-        size_t num_levels = 0;
-
-        if (_key_reader != nullptr) {
-            _key_reader->get_levels(&def_levels, &rep_levels, &num_levels);
-        } else if (_value_reader != nullptr) {
-            _value_reader->get_levels(&def_levels, &rep_levels, &num_levels);
-        } else {
-            DCHECK(false) << "Unreachable!";
-        }
-
-        auto& offsets = map_column->offsets_column()->get_data();
-        offsets.resize(num_levels + 1);
-        NullColumn null_column(num_levels);
-        auto& is_nulls = null_column.get_data();
-        size_t num_offsets = 0;
-        bool has_null = false;
-
-        // ParquetFiled Map -> Map<Struct<key,value>>
-        def_rep_to_offset(_field->level_info, def_levels, rep_levels, num_levels, &offsets[0], &is_nulls[0],
-                          &num_offsets, &has_null);
-        offsets.resize(num_offsets + 1);
-        is_nulls.resize(num_offsets);
-
-        // fill with default
-        if (_key_reader == nullptr) {
-            key_column->append_default(offsets.back());
-        }
-        if (_value_reader == nullptr) {
-            value_column->append_default(offsets.back());
-        }
-
-        if (dst->is_nullable()) {
-            DCHECK(nullable_column != nullptr);
-            nullable_column->mutable_null_column()->swap_column(null_column);
-            nullable_column->set_has_null(has_null);
-        }
-
-        return Status::OK();
-    }
-
-    void get_levels(level_t** def_levels, level_t** rep_levels, size_t* num_levels) override {
-        // check _value_reader
-        if (_key_reader != nullptr) {
-            _key_reader->get_levels(def_levels, rep_levels, num_levels);
-        } else if (_value_reader != nullptr) {
-            _value_reader->get_levels(def_levels, rep_levels, num_levels);
-        } else {
-            DCHECK(false) << "Unreachable!";
-        }
-    }
-
-    void set_need_parse_levels(bool need_parse_levels) override {
-        if (_key_reader != nullptr) {
-            _key_reader->set_need_parse_levels(need_parse_levels);
-        }
-
-        if (_value_reader != nullptr) {
-            _value_reader->set_need_parse_levels(need_parse_levels);
-        }
-    }
-
-    void collect_column_io_range(std::vector<io::SharedBufferedInputStream::IORange>* ranges, int64_t* end_offset,
-                                 ColumnIOType type, bool active) override {
-        if (_key_reader != nullptr) {
-            _key_reader->collect_column_io_range(ranges, end_offset, type, active);
-        }
-        if (_value_reader != nullptr) {
-            _value_reader->collect_column_io_range(ranges, end_offset, type, active);
-        }
-    }
-
-    void select_offset_index(const SparseRange<uint64_t>& range, const uint64_t rg_first_row) override {
-        if (_key_reader != nullptr) {
-            _key_reader->select_offset_index(range, rg_first_row);
-        }
-        if (_value_reader != nullptr) {
-            _value_reader->select_offset_index(range, rg_first_row);
-        }
-    }
-
-private:
-    const ParquetField* _field = nullptr;
-    std::unique_ptr<ColumnReader> _key_reader;
-    std::unique_ptr<ColumnReader> _value_reader;
-};
-
-class StructColumnReader : public ColumnReader {
-public:
-    explicit StructColumnReader() = default;
-    ~StructColumnReader() override = default;
-
-    Status init(const ParquetField* field, std::map<std::string, std::unique_ptr<ColumnReader>>&& child_readers) {
-        _field = field;
-        _child_readers = std::move(child_readers);
-
-        if (_child_readers.empty()) {
-            return Status::InternalError("No avaliable parquet subfield column reader in StructColumn");
-        }
-
-        for (const auto& pair : _child_readers) {
-            if (pair.second != nullptr) {
-                _def_rep_level_child_reader = &(pair.second);
-                return Status::OK();
-            }
-        }
-
-        return Status::InternalError("No existed parquet subfield column reader in StructColumn");
-    }
-
-    Status read_range(const Range<uint64_t>& range, const Filter* filter, ColumnPtr& dst) override {
-        NullableColumn* nullable_column = nullptr;
-        StructColumn* struct_column = nullptr;
-        if (dst->is_nullable()) {
-            nullable_column = down_cast<NullableColumn*>(dst.get());
-            DCHECK(nullable_column->mutable_data_column()->is_struct());
-            struct_column = down_cast<StructColumn*>(nullable_column->mutable_data_column());
-        } else {
-            DCHECK(dst->is_struct());
-            DCHECK(!_field->is_nullable);
-            struct_column = down_cast<StructColumn*>(dst.get());
-        }
-
-        const auto& field_names = struct_column->field_names();
-
-        DCHECK_EQ(field_names.size(), _child_readers.size());
-
-        // Fill data for subfield column reader
-        size_t real_read = 0;
-        bool first_read = true;
-        for (size_t i = 0; i < field_names.size(); i++) {
-            const auto& field_name = field_names[i];
-            if (LIKELY(_child_readers.find(field_name) != _child_readers.end())) {
-                if (_child_readers[field_name] != nullptr) {
-                    auto& child_column = struct_column->field_column(field_name);
-                    RETURN_IF_ERROR(_child_readers[field_name]->read_range(range, filter, child_column));
-                    real_read = child_column->size();
-                    first_read = false;
-                }
-            } else {
-                return Status::InternalError(
-                        strings::Substitute("there is no match subfield reader for $1", field_name));
-            }
-        }
-
-        if (UNLIKELY(first_read)) {
-            return Status::InternalError(
-                    strings::Substitute("All used subfield of struct type $1 is not exist", _field->name));
-        }
-
-        for (size_t i = 0; i < field_names.size(); i++) {
-            const auto& field_name = field_names[i];
-            if (_child_readers[field_name] == nullptr) {
-                Column* child_column = struct_column->field_column(field_name).get();
-                child_column->append_default(real_read);
-            }
-        }
-
-        if (dst->is_nullable()) {
-            DCHECK(nullable_column != nullptr);
-            size_t row_nums = struct_column->fields_column()[0]->size();
-            NullColumn null_column(row_nums, 0);
-            auto& is_nulls = null_column.get_data();
-            bool has_null = false;
-            _handle_null_rows(is_nulls.data(), &has_null, row_nums);
-
-            nullable_column->mutable_null_column()->swap_column(null_column);
-            nullable_column->set_has_null(has_null);
-        }
-        return Status::OK();
-    }
-
-    void set_can_lazy_decode(bool can_lazy_decode) override {
-        for (const auto& kv : _child_readers) {
-            if (kv.second == nullptr) continue;
-            kv.second->set_can_lazy_decode(can_lazy_decode);
-        }
-    }
-
-    // get_levels functions only called by complex type
-    // If parent is a struct type, only def_levels has value.
-    // If parent is list or map type, def_levels & rep_levels both have value.
-    void get_levels(level_t** def_levels, level_t** rep_levels, size_t* num_levels) override {
-        for (const auto& pair : _child_readers) {
-            // Considering not existed subfield, we will not create its ColumnReader
-            // So we should pick up the first existed subfield column reader
-            if (pair.second != nullptr) {
-                pair.second->get_levels(def_levels, rep_levels, num_levels);
-                return;
-            }
-        }
-    }
-
-    void set_need_parse_levels(bool need_parse_levels) override {
-        for (const auto& pair : _child_readers) {
-            if (pair.second != nullptr) {
-                pair.second->set_need_parse_levels(need_parse_levels);
-            }
-        }
-    }
-
-    bool try_to_use_dict_filter(ExprContext* ctx, bool is_decode_needed, const SlotId slotId,
-                                const std::vector<std::string>& sub_field_path, const size_t& layer) override {
-        if (sub_field_path.size() <= layer) {
-            return false;
-        }
-        const std::string& sub_field = sub_field_path[layer];
-        if (_child_readers.find(sub_field) == _child_readers.end()) {
-            return false;
-        }
-
-        if (_child_readers[sub_field] == nullptr) {
-            return false;
-        }
-        return _child_readers[sub_field]->try_to_use_dict_filter(ctx, is_decode_needed, slotId, sub_field_path,
-                                                                 layer + 1);
-    }
-
-    Status rewrite_conjunct_ctxs_to_predicate(bool* is_group_filtered, const std::vector<std::string>& sub_field_path,
-                                              const size_t& layer) override {
-        const std::string& sub_field = sub_field_path[layer];
-        return _child_readers[sub_field]->rewrite_conjunct_ctxs_to_predicate(is_group_filtered, sub_field_path,
-                                                                             layer + 1);
-    }
-
-    Status filter_dict_column(const ColumnPtr& column, Filter* filter, const std::vector<std::string>& sub_field_path,
-                              const size_t& layer) override {
-        const std::string& sub_field = sub_field_path[layer];
-        StructColumn* struct_column = nullptr;
-        if (column->is_nullable()) {
-            NullableColumn* nullable_column = down_cast<NullableColumn*>(column.get());
-            DCHECK(nullable_column->mutable_data_column()->is_struct());
-            struct_column = down_cast<StructColumn*>(nullable_column->mutable_data_column());
-        } else {
-            DCHECK(column->is_struct());
-            DCHECK(!_field->is_nullable);
-            struct_column = down_cast<StructColumn*>(column.get());
-        }
-        return _child_readers[sub_field]->filter_dict_column(struct_column->field_column(sub_field), filter,
-                                                             sub_field_path, layer + 1);
-    }
-
-    Status fill_dst_column(ColumnPtr& dst, const ColumnPtr& src) override {
-        StructColumn* struct_column_src = nullptr;
-        StructColumn* struct_column_dst = nullptr;
-        if (src->is_nullable()) {
-            NullableColumn* nullable_column_src = down_cast<NullableColumn*>(src.get());
-            DCHECK(nullable_column_src->mutable_data_column()->is_struct());
-            struct_column_src = down_cast<StructColumn*>(nullable_column_src->mutable_data_column());
-            NullColumn* null_column_src = nullable_column_src->mutable_null_column();
-            NullableColumn* nullable_column_dst = down_cast<NullableColumn*>(dst.get());
-            DCHECK(nullable_column_dst->mutable_data_column()->is_struct());
-            struct_column_dst = down_cast<StructColumn*>(nullable_column_dst->mutable_data_column());
-            NullColumn* null_column_dst = nullable_column_dst->mutable_null_column();
-            null_column_dst->swap_column(*null_column_src);
-            nullable_column_src->update_has_null();
-            nullable_column_dst->update_has_null();
-        } else {
-            DCHECK(src->is_struct());
-            DCHECK(dst->is_struct());
-            DCHECK(!_field->is_nullable);
-            struct_column_src = down_cast<StructColumn*>(src.get());
-            struct_column_dst = down_cast<StructColumn*>(dst.get());
-        }
-        const auto& field_names = struct_column_dst->field_names();
-        for (size_t i = 0; i < field_names.size(); i++) {
-            const auto& field_name = field_names[i];
-            if (LIKELY(_child_readers.find(field_name) != _child_readers.end())) {
-                if (_child_readers[field_name] == nullptr) {
-                    struct_column_dst->field_column(field_name)
-                            ->swap_column(*(struct_column_src->field_column(field_name)));
-                } else {
-                    RETURN_IF_ERROR(_child_readers[field_name]->fill_dst_column(
-                            struct_column_dst->field_column(field_name), struct_column_src->field_column(field_name)));
-                }
-            } else {
-                return Status::InternalError(
-                        strings::Substitute("there is no match subfield reader for $1", field_name));
-            }
-        }
-        return Status::OK();
-    }
-
-    void collect_column_io_range(std::vector<io::SharedBufferedInputStream::IORange>* ranges, int64_t* end_offset,
-                                 ColumnIOType type, bool active) override {
-        for (const auto& pair : _child_readers) {
-            if (pair.second != nullptr) {
-                pair.second->collect_column_io_range(ranges, end_offset, type, active);
-            }
-        }
-    }
-
-    void select_offset_index(const SparseRange<uint64_t>& range, const uint64_t rg_first_row) override {
-        for (const auto& pair : _child_readers) {
-            if (pair.second != nullptr) {
-                pair.second->select_offset_index(range, rg_first_row);
-            }
-        }
-    }
-
-private:
-    void _handle_null_rows(uint8_t* is_nulls, bool* has_null, size_t num_rows) {
-        level_t* def_levels = nullptr;
-        level_t* rep_levels = nullptr;
-        size_t num_levels = 0;
-        (*_def_rep_level_child_reader)->get_levels(&def_levels, &rep_levels, &num_levels);
-
-        if (def_levels == nullptr) {
-            // If subfields are required, def_levels is nullptr
-            *has_null = false;
-            return;
-        }
-
-        LevelInfo level_info = _field->level_info;
-
-        if (rep_levels != nullptr) {
-            // It's a RepeatedStoredColumnReader
-            size_t rows = 0;
-            for (size_t i = 0; i < num_levels; i++) {
-                if (def_levels[i] < level_info.immediate_repeated_ancestor_def_level ||
-                    rep_levels[i] > level_info.max_rep_level) {
-                    continue;
-                }
-
-                // Start for a new row
-                if (def_levels[i] >= level_info.max_def_level) {
-                    is_nulls[rows] = 0;
-                } else {
-                    is_nulls[rows] = 1;
-                    *has_null = true;
-                }
-                rows++;
-            }
-            DCHECK_EQ(num_rows, rows);
-        } else {
-            // For OptionalStoredColumnReader, num_levels is equal to num_rows
-            DCHECK(num_rows == num_levels);
-            for (size_t i = 0; i < num_levels; i++) {
-                if (def_levels[i] >= level_info.max_def_level) {
-                    is_nulls[i] = 0;
-                } else {
-                    is_nulls[i] = 1;
-                    *has_null = true;
-                }
-            }
-        }
-    }
-
-    // _field is generated by parquet format, so it's child order may different from _child_readers.
-    const ParquetField* _field = nullptr;
-    // _children_readers order is the same as TypeDescriptor children order.
-    std::map<std::string, std::unique_ptr<ColumnReader>> _child_readers;
-    // First non-nullptr child ColumnReader, used to get def & rep levels
-    const std::unique_ptr<ColumnReader>* _def_rep_level_child_reader = nullptr;
-};
 
 void ColumnReader::get_subfield_pos_with_pruned_type(const ParquetField& field, const TypeDescriptor& col_type,
                                                      bool case_sensitive, std::vector<int32_t>& pos) {
